Back-end pronto! Front-end em desenvolvimento.

+++ aquilax/src/main/java/br/com/mackenzie/aquilax/model/Drone.java
@@ -1,6 +1,11 @@
 package br.com.mackenzie.aquilax.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "drone")
@@ -9,24 +14,22 @@ public class Drone {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_drone")
-    private Long idDrone;
+    private Integer idDrone;
 
     private String modelo;
 
     @Column(name = "num_identificacao")
     private int numIdentificacao;
 
-    private int status; // 1 = Pronto, 0 = Manutenção
+    private int status;
 
     @Column(name = "id_frota")
     private int idFrota;
-
-    // Construtor padrão (Obrigatório)
     public Drone() {}
 
     // Getters e Setters (Para o Spring conseguir ler os dados)
-    public Long getIdDrone() { return idDrone; }
-    public void setIdDrone(Long idDrone) { this.idDrone = idDrone; }
+    public Integer getIdDrone() { return idDrone; }
+    public void setIdDrone(Integer idDrone) { this.idDrone = idDrone; }
 
     public String getModelo() { return modelo; }
     public void setModelo(String modelo) { this.modelo = modelo; }
